backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
SQLAlchemy==2.0.32
pydantic==2.8.2
pydantic-settings==2.4.0
python-multipart==0.0.9
PyJWT==2.9.0
psycopg[binary]==3.2.1
google-genai==1.73.1
